requirements.txt
flask>=3.0,<4.0
python-socketio>=5.11,<6.0
python-engineio>=4.9,<5.0
pika>=1.3,<2.0
PyJWT>=2.8,<3.0
cryptography>=42.0,<44.0
gunicorn>=22.0,<23.0
requests>=2.31,<3.0
